requirements.txt
path.py
plotly
scikit-learn
matplotlib
seaborn

torchvision==0.10.0+cu102
torchaudio===0.9.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.9.0+cu102